GlobalState in OpState, move check_unstable
