+++ cli/main.rs
@@ -78,9 +78,9 @@ use flags::DenoSubcommand;
 use flags::Flags;
 use futures::future::FutureExt;
 use futures::Future;
+use global_state::exit_unstable;
 use log::Level;
 use log::LevelFilter;
-use state::exit_unstable;
 use std::env;
 use std::io::Read;
 use std::io::Write;
