+++ cli/ops/runtime_compiler.rs
@@ -36,7 +36,7 @@ async fn op_compile(
   let cli_state = super::cli_state2(&state);
   cli_state.check_unstable("Deno.compile");
   let args: CompileArgs = serde_json::from_value(args)?;
-  let global_state = cli_state.global_state.clone();
+  let global_state = cli_state.clone();
   let permissions = {
     let state = state.borrow();
     state.borrow::<Permissions>().clone()
@@ -78,7 +78,7 @@ async fn op_transpile(
   let cli_state = super::cli_state2(&state);
   cli_state.check_unstable("Deno.transpile");
   let args: TranspileArgs = serde_json::from_value(args)?;
-  let global_state = cli_state.global_state.clone();
+  let global_state = cli_state.clone();
   let permissions = {
     let state = state.borrow();
     state.borrow::<Permissions>().clone()
